chore(server): core IoC #83 - (most) commits resolvers to TS
