+++ packages/server/modules/core/domain/commits/operations.ts
@@ -15,7 +15,12 @@ import {
 } from '@/modules/core/graph/generated/graphql'
 import { BranchCommitRecord, StreamCommitRecord } from '@/modules/core/helpers/types'
 import { BatchedSelectOptions } from '@/modules/shared/helpers/dbHelper'
-import { MaybeNullOrUndefined, Nullable, Optional } from '@speckle/shared'
+import {
+  MaybeNullOrUndefined,
+  Nullable,
+  NullableKeysToOptional,
+  Optional
+} from '@speckle/shared'
 import { Knex } from 'knex'
 
 export type GetCommits = (
@@ -47,13 +52,11 @@ export type GetSpecificBranchCommits = (
 ) => Promise<CommitWithBranchId[]>
 
 export type StoreCommit = (
-  params: Omit<Commit, 'id' | 'createdAt'> & {
-    message?: Nullable<string>
-  }
+  params: Omit<NullableKeysToOptional<Commit>, 'id' | 'createdAt'>
 ) => Promise<Commit>
 
 export type CreateCommitByBranchId = (
-  params: {
+  params: NullableKeysToOptional<{
     streamId: string
     branchId: string
     objectId: string
@@ -62,14 +65,14 @@ export type CreateCommitByBranchId = (
     sourceApplication: Nullable<string>
     totalChildrenCount?: MaybeNullOrUndefined<number>
     parents: Nullable<string[]>
-  },
+  }>,
   options?: Partial<{
     notify: boolean
   }>
 ) => Promise<Commit>
 
 export type CreateCommitByBranchName = (
-  params: {
+  params: NullableKeysToOptional<{
     streamId: string
     branchName: string
     objectId: string
@@ -78,7 +81,7 @@ export type CreateCommitByBranchName = (
     sourceApplication: Nullable<string>
     totalChildrenCount?: MaybeNullOrUndefined<number>
     parents: Nullable<string[]>
-  },
+  }>,
   options?: Partial<{
     notify: boolean
   }>
